Refactor git commands (add mutating commands)
